+++ src/main/java/sayTheSpire/UIPatches/monsters/AbstractMonsterPatch.java
@@ -5,4 +5,11 @@ import sayTheSpire.Output;
 
 public class AbstractMonsterPatch {
 
+    @SpirePatch(clz = AbstractMonster.class, method = "die", paramtypez = { boolean.class })
+    public static class DiePatch {
+
+        public static void Prefix(AbstractMonster __instance, boolean triggerRelics) {
+            Output.event(new TextEvent(__instance.name + " defeated"));
+        }
+    }
 }