+++ docs_src/src/changes.md
@@ -6,6 +6,7 @@ This should be the last general bug fix and gameplay notification patch before m
 * Added: Achievement items on the character stats screen now have location information.
 * Fixed: Block is now correctly read out if block was already greater than 0 (it was previously triggering on the wrong event)
 * Added: Combat orb events are now read out (such as when they are channeled and when they are evoked). This can be configured; see the config section of the documentation for more info.
+* Added: Monster deaths are now read out.
 
 ## 0.3.2-beta